+++ g2p_programs/models/accounting/__init__.py
@@ -17,15 +17,4 @@
 # limitations under the License.
 #
 
-from . import constants
-from . import job_related_mixin
-from . import managers
-from . import programs
-from . import registrant
-from . import registrant_attribute
-from . import program_membership
-from . import cycle
-from . import cycle_membership
-from . import voucher
-from . import cash_voucher
-from . import accounting
+from . import account_journal